fast-follow sweep: the five carried findings from M16's final review
Five small items, none behavioural except where a test was proving less than it claimed.

Two were real. The 12a/12b thread-oracle tests asserted only `rep.code == 3` — the exit
convention for EVERY divergence — so a trace that diverged for an unrelated reason would
have passed while proving nothing about the thread oracle. They now pin the message
`verify_thread` emits; mutating that message fails four of the five tests in the file, and
the fifth correctly survives because the delivery landmark uses `mirror_delivery`'s own
inline check with different wording. And 12b's claim to be "stronger than 12a" — the
sigreturn tag is a nonzero child id — lived only in its doc comment, so a fixture whose
schedule moved sigreturn onto main would have silently degraded it into a duplicate of
12a. That precondition is now asserted.

C7 was the same class: `catch_unwind(...).is_err()` proved only that SOMETHING unwound,
and `deliver_signal_to` has since grown a second way to panic (`check_deliverable`). It
now downcasts the payload and pins the nested-delivery guard's own text.

The rest are cosmetic. `dbg_regs_of` takes `usize` like its sibling `dbg_kport_of` and
like `Box_`, with the one cast moved to the CLI boundary where the u32 originates.
`on_altstack`'s doc said "unchanged for every existing caller", a migration note that
outlived its migration, and now says what the function means. Record's fault arm calls
`deliver_signal_to(thread, …)` rather than the `deliver_signal(…)` shorthand it delegates
to, so the pair is grep-identical with replay's mirror — rule 1 held either way, but this
branch works to make symmetry visible rather than inferable.

+++ crates/retrace/src/debug.rs
@@ -352,7 +352,9 @@ impl<'a> Exec<'a> {
     /// becomes an `Err` here (a usage error the CLI reports and exits 5 on, same as `break`'s 6-slot
     /// limit), never a panic on an unwrap.
     fn cmd_regs_of<W: Write>(&mut self, tid: u32, out: &mut W) -> Result<(), String> {
-        match self.sess().dbg_regs_of(tid) {
+        // `tid` is u32 because that is what the CLI parses; the session's hooks take `usize`,
+        // matching `Box_`. One cast, at the boundary where the u32 actually comes from.
+        match self.sess().dbg_regs_of(tid as usize) {
             Some(dump) => line(out, format_args!("{dump}")),
             None => Err(format!("no such thread: {tid}")),
         }
